Make review bounds state global and bound review by default where supported.
